Merilife/MediBot/chatbot-website/requirements.txt
Django==5.1
mysqlclient==2.1.1
djangorestframework==3.14.0
# numpy==1.23.5
# pandas==1.5.3
# transformers==4.21.1
gunicorn==20.1.0
django-cors-headers==3.14.0
sounddevice==0.4.6
soundfile==0.12.1
# scipy==1.10.1
django-environ==0.12.0
python-dotenv==1.0.0
django-allauth==0.60.0
django-guardian==2.4.0
pymongo==4.6.1
google-genai==1.11.0
SpeechRecognition==3.10.0
pydub==0.25.1
reportlab==4.0.9

# Testing
pytest-django==4.5.2
coverage==7.4.4
selenium==4.20.0
behave-django==1.3.0
allpairspy==2.5.0
# WebDriver for Selenium
webdriver-manager==4.0.1
ffmpeg-python==0.2.0
factory_boy
faker
pytest-cov